% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/markedPP_funcs.R
\name{calc.firstOrd}
\alias{calc.firstOrd}
\title{Calculate first order characteristics from Illian et al.}
\usage{
calc.firstOrd(x, W)
}
\arguments{
\item{x}{Data.frame of \code{<t, m>} single pair of event series.}

\item{W}{Observation window \code{c(low, high)}, equivalent to \code{xlim}.}
}
\value{
List containing the following:
  \itemize{
    \item \code{lambda}: vector of intensities of each process and both combined
    \item \code{p}: matrix of joint probs of nearest neighbor marks, \code{p_11, p_12, p_21, p_22}
    \item \code{p_dot}: vector of margin values (columns) of p, \code{c(p_dot1, p_dot2)}
    \item \code{markp}: vector of mark probabilities, \code{c(p_1,p_2)}
    \item \code{S}: coefficient of segregation
    \item \code{M}: vector of mingling index with \code{k=1} for observed process and theoretical
      value if indep, respectively, \code{c(M_k, M_ki)}
  }
}
\description{
Calculate first order characteristics from Illian et al.
}
