% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iet_funcs.R
\name{calc.forward_inter_event_times}
\alias{calc.forward_inter_event_times}
\title{Calculate inter-event times between one event and the next event forward in time.}
\usage{
calc.forward_inter_event_times(x)
}
\arguments{
\item{x}{Data.frame of \code{<id, mark, time>} for pair of event series.}
}
\value{
List of lists:
  \itemize{
    \item \code{AB} = set of forward iets for fixed mark 1 time to closest event of mark 2
    \item \code{BA} = set of forward iets for fixed mark 2 time to closest event of mark 1
  }
}
\description{
\code{calc.forward_inter_event_times} calculates inter-event times between one
event and the next event forward in time. (i.e. equiv to inter-event time in
Poisson process).
}
