% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iet_funcs.R
\name{calc.iet_SLR}
\alias{calc.iet_SLR}
\title{Calculate score-based likelihood ratios.}
\usage{
calc.iet_SLR(ids, within, between, bds.mn, bds.md, match = TRUE,
  bw.type = "nrd")
}
\arguments{
\item{ids}{Vector of unique ids of the data considered.}

\item{within}{Data.frame of inter-event times for same-source pairs of event series,
\code{<id, mn, md>}}

\item{between}{Data.frame of inter-event times for different-source pairs of event
series, \code{<id.A, id.B, mn, md>}}

\item{bds.mn, bds.md}{Numeric array \code{c(lower, upper)} of limmits for KDE on
mean/median IET.}

\item{match}{Logical indicating to compute scores of known matches or non-matches;
\code{default == TRUE}}

\item{bw.type}{Bandwidth estimation method to use in density estimation
one of \code{c("nrd", "nrd0", "ucv", "bcv", "SJ")}; \code{default == "nrd"}}
}
\value{
Data.frame of score-based likelihood ratios:
  \itemize{
    \item \code{match == TRUE}: \code{<id, mn, md>}
    \item \code{match == FALSE}: \code{<id.A, id.B, mn, md>}
  }
}
\description{
Calculate score-based likelihood ratios.
}
