% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iet_funcs.R
\name{calc.inter_event_times}
\alias{calc.inter_event_times}
\title{Calculate inter-event times for a pair of event series.}
\usage{
calc.inter_event_times(x, sum.stat = FALSE)
}
\arguments{
\item{x}{Data.frame of \code{<id, mark, time>} for pair of event series.}
}
\value{
List of lists:
  \itemize{
    \item \code{AB} = set of iets for fixed mark 1 time to closest event of mark 2
    \item \cdoe{BA} = set of iets for fixed mark 2 time to closest event of mark 1
  }
}
\description{
\code{calc.inter_event_times} calculates inter-event times regardless of direction
(i.e. all positive even if closest event was earlier).
}
