% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/markedPP_funcs.R
\name{calc.SLR}
\alias{calc.SLR}
\title{Calculate score-based likelihood ratios.}
\usage{
calc.SLR(ids, within, between, match = TRUE, bw.type = "nrd",
  s.bds = c(-1, 1), m.bds = c(0, 0.6))
}
\arguments{
\item{ids}{Vector of unique ids of the data considered.}

\item{within}{Data.frame of inter-event times for same-source pairs of event series,
\code{<id, mn, md>}}

\item{between}{Data.frame of inter-event times for different-source pairs of event
series, \code{<id.A, id.B, mn, md>}}

\item{match}{Logical indicating to compute scores of known matches or non-matches;
\code{default == TRUE}}

\item{bw.type}{Bandwidth estimation method to use in density estimation
one of \code{c("nrd", "nrd0", "ucv", "bcv", "SJ")}; \code{default == "nrd"}}

\item{s.bds, m.bds}{Bounds on density estimation of segregation and mingling, respectively;
\code{default == c(-1,1), default == c(0, 0.6)}}
}
\value{
Data.frame of score-based likelihood ratios:
  \itemize{
    \item \code{match == TRUE}: \code{<id, s, m>}
    \item \code{match == FALSE}: \code{<id.A, id.B, s, m>}
  }
}
\description{
Calculate score-based likelihood ratios.
}
