% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score_funcs.R
\name{calc_cmp}
\alias{calc_cmp}
\title{Calculate the coincidental match probability for a single pair of event
series.}
\usage{
calc_cmp(data, n, W = c(0, 7), bidirectional = TRUE, samp = "empirical",
  rng = NULL)
}
\arguments{
\item{data}{List of data.frames for one pair of event series (output of
\code{assocr::sessionize_data()}):
\itemize{
  \item \code{data}: sessionized data, \code{<id, m, sid, t>}
  \item \code{sessions}: summary of sessionized data,
        \code{<id, m, sid, n, t>}
}}

\item{n}{Number of samples to draw}

\item{W}{Vector containing observation window \code{c(low, high)}.}

\item{bidirectional}{Logical indicating if the inter-event times should be
bidirectional (versus only looking at events prior to event of interest)}

\item{samp}{Sampling technique to use. One of
\itemize{
  \item \code{empirical}: sessionized resampling from empirical distn of
    start times
  \item \code{gaussian}: sessionized resampling from Gaussian distn such
    that 99% of start times fall in \code{rng}
  \item \code{uniform}: sessionized resampling from Uniform distn over
    \code{rng}
}}

\item{rng}{Vector of \code{c(low, high)} limits for sampling session start
times if \code{samp == "periodic"}}
}
\value{
Data.frame of CMPs for each score function;
  \code{<iet.mn, iet.md, s, m, cmp.iet.mn, cmp.iet.md, cmp.s, cmp.m>}
}
\description{
Calculate the coincidental match probability for a single pair of event
series.
}
