% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score_funcs.R
\name{calc_score_funcs}
\alias{calc_score_funcs}
\title{Calculate the coefficient of segregation, mingling index, and inter-event
time summary statistics of a given pair of event series.}
\usage{
calc_score_funcs(x, W = c(0, 7), bidirectional = TRUE)
}
\arguments{
\item{x}{Data.frame of \code{<m, t>} for a single pair of event series.}

\item{W}{Vector containing observation window \code{c(low, high)}.}

\item{bidirectional}{Logical indicating if the inter-event times should be
bidirectional (versus only looking at events prior to event of interest)}
}
\value{
List containing
  \itemize{
    \item \code{iet.mn}: mean iet for fixed mark 2/B to mark 1/A
    \item \code{iet.md}: median iet for fixed mark 2/B to mark 1/A
    \item \code{s}: coefficient of segregation
    \item \code{m}: mingling index, k=1
    \item \code{nA}: number of events of type 1/A
    \item \code{nB}: number of events of type 2/B
  }
}
\description{
Calculate the coefficient of segregation, mingling index, and inter-event
time summary statistics of a given pair of event series.
}
