% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score_funcs.R
\name{calc_slr}
\alias{calc_slr}
\title{Calculate score-based likelihood ratios for pairs of event series.}
\usage{
calc_slr(same_src, diff_src, bds.iet.mn = c(0, 1), bds.iet.md = c(0, 1),
  bds.s = c(-1, 1), bds.m = c(0, 0.6), bw.type = "nrd",
  loo_xval = FALSE)
}
\arguments{
\item{same_src, diff_src}{Data.frame of score functions for same- and
different-source pairs of event series, respectively, computed via
\code{assocr::calc_score_funcs}; i.e., \code{<iet.mn, iet.md, s, m>}}

\item{bw.type}{Bandwidth estimation method to use in density estimation,
one of \code{c("nrd", "nrd0", "ucv", "bcv", "SJ")}}

\item{loo_xval}{Boolean indicating whether or not to use leave-pairs-out
cross-validation for computing the SLR and CMP (default FALSE). If TRUE,
must have id.A and id.B columns in both same_src and diff_src dataframes}

\item{bds.iet, bds.s, bds.m}{Vector of bounds \code{c(low, high)} for density
estimation of mean/median inter-event times, segregation and mingling,
respectively.}
}
\value{
Data.frame of SLRs and CMPs for each score function where
  \code{indep == 1} if the pair was independent; i.e.,
  \code{<iet.mn, iet.md, s, m, indep, slr.iet.mn, slr.iet.md, slr.s, slr.m,
         cmp.iet.mn, cmp.iet.md, cmp.s, cmp.m>}
}
\description{
Calculate score-based likelihood ratios for pairs of event series.
}
