% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score_funcs.R
\name{calc_slr_cmp}
\alias{calc_slr_cmp}
\title{Calculate score-based likelihood ratios & coincidental match probabilities
for simulated pairs of event series.}
\usage{
calc_slr_cmp(same_src, diff_src, bds.iet, bds.s = c(-1, 1), bds.m = c(0,
  0.6), bw.type = "nrd")
}
\arguments{
\item{same_src, diff_src}{Data.frame of score functions for same- and
different-source pairs of event series, respectively, computed via
\code{assocr::calc_score_funcs}; i.e., \code{<iet.mn, iet.md, s, m>}}

\item{bds.iet, bds.s, bds.m}{Vector of bounds \code{c(low, high)} for density
estimation of mean/median inter-event times, segregation and mingling,
respectively.}

\item{bw.type}{Bandwidth estimation method to use in density estimation,
one of \code{c("nrd", "nrd0", "ucv", "bcv", "SJ")}}
}
\value{
Data.frame of SLRs and CMPs for each score function where
  \code{indep == 1} if the pair was independent; i.e.,
  \code{<iet.mn, iet.md, s, m, indep, slr.iet.mn, slr.iet.md, slr.s, slr.m,
         cmp.iet.mn, cmp.iet.md, cmp.s, cmp.m>}
}
\description{
Calculate score-based likelihood ratios & coincidental match probabilities
for simulated pairs of event series.
}
