% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score_funcs.R
\name{ceiling_dec}
\alias{ceiling_dec}
\title{Return the ceiling of a decimal number as a decimal with user-specified
precision.}
\usage{
ceiling_dec(x, level = 0)
}
\arguments{
\item{x}{Positive double.}

\item{level}{Number of decimal places to return; \code{default == 0}}
}
\value{
Decimal ceiling of \code{x}.
}
\description{
Return the ceiling of a decimal number as a decimal with user-specified
precision.
}
