% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score_funcs.R
\name{count_zeros}
\alias{count_zeros}
\title{Count the number of leading zeros between decimal and first non-zero entry.}
\usage{
count_zeros(x)
}
\arguments{
\item{x}{Positive double.}
}
\value{
Number of leading zeros.
}
\description{
Count the number of leading zeros between decimal and first non-zero entry.
}
