% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score_funcs.R
\name{filter_list}
\alias{filter_list}
\title{Filter list of data.frames to contain one pair of event series.}
\usage{
filter_list(data, id, m = NULL)
}
\arguments{
\item{data}{List of data.frames for two pairs of event series (output of
\code{assocr::sessionize_data()}):
\itemize{
  \item \code{data}: sessionized data, \code{<id, m, sid, t>}
  \item \code{sessions}: summary of sessionized data,
        \code{<id, m, sid, n, t>}
}}

\item{id}{Either of the following
\itemize{
  \item String containing ID of user for same-source pair of event streams
  \item Vector of strings containing IDs for different-source pair of
    event streams
}}

\item{m}{Array of c(mark of id 1, mark of id 2) used for different-source
pairs of event streams if \code{id == c(id1, id2)}}
}
\description{
Filter list of data.frames to contain one pair of event series.
}
