% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/markedPP_funcs.R
\name{find_sessions}
\alias{find_sessions}
\title{Group single event series (i.e., one user & one type of event) into sessions.}
\usage{
find_sessions(data, thres = 10, timeScale = "min")
}
\arguments{
\item{data}{Data.frame of event data of same mark for one user,
\code{<id, t>}}

\item{thres}{Number of inactive minutes to consider a new session start}

\item{timeScale}{Scale of time measurements of data; one of
\code{c("s", "min", "hr", "day")}}
}
\value{
List of data.frames:
  \itemize{
    \item \code{data}: sessionized data, \code{<t, id, t.min, diff, sid>}
    \item \code{sessions}: summary of sessions, \code{<id, t, t.min, sid, n>}
  }
}
\description{
Group single event series (i.e., one user & one type of event) into sessions.
}
