% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iet_funcs.R
\name{my.min}
\alias{my.min}
\title{Return smallest non-zero element of a vector.}
\usage{
my.min(x)
}
\arguments{
\item{x}{A vector of positive real numbers.}
}
\value{
Minimum value of \code{x} such that \code{min(x) > 0}
}
\description{
Return smallest non-zero element of a vector.
}
\examples{
my.min(c(1,3,5,2,0))
}
