% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score_funcs.R
\name{nonzero.min}
\alias{nonzero.min}
\title{Return smallest non-zero element of a vector.}
\usage{
nonzero.min(x)
}
\arguments{
\item{x}{A vector of positive real numbers.}
}
\value{
Minimum value of \code{x} such that \code{min(x) > 0}
}
\description{
Return smallest non-zero element of a vector.
}
\examples{
nonzero.min(c(1,3,5,2,0))
}
