% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iet_funcs.R
\name{plot.iet}
\alias{plot.iet}
\title{Plot inter event time empirical cdf's for all pairwise combinations of event series.}
\usage{
\method{plot}{iet}(x, main = "", sub = "", samp.prob = 1, xmax = NULL,
  legend = FALSE)
}
\arguments{
\item{x}{Data.frame of inter-event times \code{<id.A, id.B, iet>}}

\item{main}{Title of the plot.}

\item{sub}{Subtitle of the plot.}

\item{samp.prob}{Probability of including diff source line to reduce overplottting; \code{deafult == 1}}

\item{xmax}{Max value for x-axis range; \code{default == max(x$iet)}}

\item{legend}{Logical indicating whether or not to include legend; \code{default == FALSE}}
}
\description{
Plot inter event time empirical cdf's for all pairwise combinations of event series.
}
