% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iet_funcs.R
\name{plot.iet_dist}
\alias{plot.iet_dist}
\title{Plot densities for a given iet statistic and pair of event streams.}
\usage{
\method{plot}{iet_dist}(within, between, which.stat = "mn",
  which.iet = "BA", bw.type = "nrd", font.sz = 12)
}
\arguments{
\item{within, between}{Data.frame of statistics returned from repeatedly running
calc.inter_event_times and saving results with same/different user event streams.}

\item{which.stat}{Mean (\code{"mn"}) or median (\code{"md"}); \code{default == "mn"}}

\item{which.iet}{Which distribution to use (\code{"AB", "BA"}); \code{default == "BA"}}

\item{bw.type}{Bandwidth selection method for density() function; \code{default == "nrd"}}

\item{font.sz}{Size of font for axis; \code{default == 16}}
}
\description{
Plot densities for a given iet statistic and pair of event streams.
}
