% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/markedPP_funcs.R
\name{plot_markedpp}
\alias{plot_markedpp}
\title{Plot single pair of event series with cumulative counts.}
\usage{
plot_markedpp(x, W, c = 1, lbl = c("Mark 1", "Mark 2"), title = "",
  xlab = "Time")
}
\arguments{
\item{x}{Data.frame of \code{<t, m>} single pair of event series.}

\item{W}{Observation window \code{c(low, high)}, equivalent to \code{xlim}.}

\item{c}{Numeric cex value for axis & legend magnification; \code{default == 1}}

\item{lbl}{Vector of labels for the legend corresponding to marks 1/A and 2/B, respectively;
\code{default == c("Mark 1", "Mark 2")}}

\item{title}{Main title of plot; \code{default == ""}}

\item{xlab}{X-axis label; \code{default == "Time"}}
}
\description{
Plot single pair of event series with cumulative counts.
}
