% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score_funcs.R
\name{session_resampling}
\alias{session_resampling}
\title{Resample session start times.}
\usage{
session_resampling(data, samp = "empirical", rng = NULL)
}
\arguments{
\item{data}{List of data.frames for one pair of event series (output of
\code{assocr::sessionize_data()}):
\itemize{
  \item \code{data}: sessionized data, \code{<id, m, sid, t>}
  \item \code{sessions}: summary of sessionized data,
        \code{<id, m, sid, n, t>}
}}

\item{samp}{Sampling technique to use. One of
\itemize{
  \item \code{empirical}: sessionized resampling from empirical distn of
    start times
  \item \code{gaussian}: sessionized resampling from Gaussian distn such
    that 99% of start times fall in \code{rng}
  \item \code{uniform}: sessionized resampling from Uniform distn over
    \code{rng}
}}

\item{rng}{Vector of \code{c(low, high)} limits for sampling session start
times if \code{samp == "periodic"}}
}
\value{
Vector of resampled times for event series of mark 1.
}
\description{
Resample session start times.
}
