% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/markedPP_funcs.R
\name{sessionize_data}
\alias{sessionize_data}
\title{Group pairs of event series data into sessions.}
\usage{
sessionize_data(data, thres = 10, timeScale = "min")
}
\arguments{
\item{data}{Data.frame of event data, \code{<id, t, m>}}

\item{thres}{Number of inactive minutes to consider a new session start}

\item{timeScale}{Scale of time measurements of data; one of
\code{c("s", "min", "hr", "day")}}
}
\value{
List of data.frames:
  \itemize{
    \item \code{data}: sessionized data, \code{<id, m, sid, t>}
    \item \code{sessions}: summary of sessionized data,
           \code{<id, m, sid, n, t>}
  }
}
\description{
Group pairs of event series data into sessions.
}
