% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score_funcs.R
\name{pairwise_cmp}
\alias{pairwise_cmp}
\title{Wrapper to compute CMP for each combination of a given pair of user data}
\usage{
pairwise_cmp(data, n, W = c(0, 7), bidirectional = TRUE,
  samp = "empirical", rng = NULL)
}
\arguments{
\item{data}{List of data.frames for two pairs of event series (output of
\code{assocr::sessionize_data()}):
\itemize{
  \item \code{data}: sessionized data, \code{<id, m, sid, t>}
  \item \code{sessions}: summary of sessionized data,
        \code{<id, m, sid, n, t>}
}}

\item{n}{Number of samples to draw}

\item{W}{Vector containing observation window \code{c(low, high)}.}

\item{bidirectional}{Logical indicating if the inter-event times should be
bidirectional (versus only looking at events prior to event of interest)}

\item{samp}{Sampling technique to use. One of
\itemize{
  \item \code{empirical}: sessionized resampling from empirical distn of
    start times
  \item \code{periodic}: sessionized resampling from sinusoidal curve
    fitted to the data
}}

\item{rng}{Vector of \code{c(low, high)} limits for sampling session start
times if \code{samp == "periodic"}}
}
\value{
Array of CMPs for each pairwise combination of event streams &
  score function
}
\description{
Wrapper to compute CMP for each combination of a given pair of user data
}
