% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/markedPP_funcs.R
\name{plot_mult_markedpp}
\alias{plot_mult_markedpp}
\title{Plot multiple pairs of event series.}
\usage{
plot_mult_markedpp(x, W, c = 1, leg = TRUE, lbl = c("Mark 1", "Mark 2"),
  leg.loc, cc = 1)
}
\arguments{
\item{x}{Data.frame of \code{<id, t, m>} multiple pairs of event series.}

\item{W}{Observation window \code{c(low, high)}, equivalent to \code{xlim}.}

\item{c}{Numeric cex value for point size; \code{default == 1}}

\item{leg}{Logical indicating whether or not to include legend; \code{default == TRUE}}

\item{lbl}{Vector of labels for the legend corresponding to marks 1/A and 2/B,
respectively; \code{default == c("Mark 1", "Mark 2")}}

\item{leg.loc}{Where to locate legend.}

\item{cc}{Numeric cex value for axis & legend magnification; \code{default == 1}}
}
\description{
Plot multiple pairs of event series.
}
